+++ nest_lesson_3/src/app.module.ts
@@ -1,10 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UsersController } from './features/users/controller/users.controller';
 import { MongooseModule } from '@nestjs/mongoose';
-import { User, UserSchema } from './features/users/domain/user.entity';
-import { UsersService } from './features/users/application/users.service';
-import { UsersQueryRepository } from './features/users/repositories/users.repository.query';
-import { UsersCommandsRepository } from './features/users/repositories/users.repository.commands';
 import { BlogsQueryRepository } from './features/blogs/repositories/blogs.repository.query';
 import { BlogsController } from './features/blogs/controller/blogs.controller';
 import { Blog, BlogSchema } from './features/blogs/domain/blog.entity';
@@ -15,23 +10,10 @@ import { PostsQueryRepository } from './features/posts/repositories/posts.reposi
 import { PostsCommandsRepository } from './features/posts/repositories/posts.repository.commands';
 import { PostsController } from './features/posts/controller/posts.controller';
 import { Post, PostSchema } from './features/posts/domain/post.entity';
-import { TestingController } from './features/testing/controller/testing.controller';
 import { appSettings } from './settings/appSettings';
-import { AuthController } from './features/auth/controller/auth.controller';
-import { AuthService } from './features/auth/application/auth.service';
-import { AuthCommandsRepository } from './features/auth/repositories/auth.repository.commands';
 import { MailerModule } from '@nestjs-modules/mailer';
 import { EmailService } from './features/mail/application/mail.service';
-import { IsConfirmationCodeActiveConstraint } from './common/IsConfirmationCodeActive';
-import { IsUserByConfirmationCodeExistConstraint } from './common/IsUserByConfirmationCodeExist';
-import { IsEmailNotConfirmedConstraint } from './common/IsEmailNotConfirmed';
-import { IsUserNotExistConstraint } from './common/IsUserNotExist';
-import { IsUserAlreadyExistConstraint } from './common/IsUserAlreadyExist';
-import { IsUserByRecoveryCodeExistConstraint } from './common/IsUserByRecoveryCodeExist';
-import { IsPasswordRecoveryCodeUsedConstraint } from './common/IsPasswordRecoveryCodeUsed';
-import { LocalStrategy } from './features/auth/application/auth.local.strategy';
 import { JwtModule } from '@nestjs/jwt';
-import { JwtStrategy } from './features/auth/application/auth.jwt.strategy';
 import { CommentsService } from './features/comments/application/comments.service';
 import {
   Comment,
@@ -46,6 +28,8 @@ import { LikesCommandsRepository } from './features/likes/repositories/likes.rep
 import { LikesQueryRepository } from './features/likes/repositories/likes.repository.query';
 import { Like, LikeSchema } from './features/likes/domain/like.entity';
 import { IsBlogExistConstraint } from './common/IsBlogExist';
+import { UsersModule } from './modules/usersModule/users.module';
+import { CommonModule } from './modules/commonModule/common.module';
 
 @Module({
   imports: [
@@ -64,10 +48,6 @@ import { IsBlogExistConstraint } from './common/IsBlogExist';
       },
     }),
     MongooseModule.forFeature([
-      {
-        name: User.name,
-        schema: UserSchema,
-      },
       { name: Blog.name, schema: BlogSchema },
       { name: Post.name, schema: PostSchema },
       { name: Comment.name, schema: CommentSchema },
@@ -78,20 +58,16 @@ import { IsBlogExistConstraint } from './common/IsBlogExist';
       secret: appSettings.api.SECRET_JWT_KEY,
       signOptions: { expiresIn: '5m' },
     }),
+    UsersModule,
+    CommonModule,
   ],
   controllers: [
-    UsersController,
     BlogsController,
     PostsController,
-    TestingController,
-    AuthController,
+    //    TestingController,
     CommentsController,
   ],
   providers: [
-    UsersService,
-    UsersQueryRepository,
-    UsersCommandsRepository,
-
     BlogsService,
     BlogsQueryRepository,
     BlogsCommandsRepository,
@@ -100,20 +76,8 @@ import { IsBlogExistConstraint } from './common/IsBlogExist';
     PostsQueryRepository,
     PostsCommandsRepository,
 
-    IsUserNotExistConstraint,
-    IsConfirmationCodeActiveConstraint,
-    IsUserByConfirmationCodeExistConstraint,
-    IsEmailNotConfirmedConstraint,
-    IsUserAlreadyExistConstraint,
-    IsUserByRecoveryCodeExistConstraint,
-    IsPasswordRecoveryCodeUsedConstraint,
     IsCommentExistConstraint,
     IsBlogExistConstraint,
-
-    AuthService,
-    LocalStrategy,
-    JwtStrategy,
-    AuthCommandsRepository,
 
     EmailService,
