+++ nest_lesson_3/src/features/comments/controller/comments.controller.ts
@@ -15,10 +15,10 @@ import {
 import { CommentsQueryRepository } from '../repositories/comments.repository.query';
 import { IsNotEmpty, IsEnum, Length } from 'class-validator';
 import { CommentsService } from '../application/comments.service';
-import { JwtAuthGuard } from 'src/features/auth/application/jwt.auth.guard';
 import { LIKE_STATUS } from 'src/common/enums';
 import { LikesQueryRepository } from 'src/features/likes/repositories/likes.repository.query';
 import { JwtService } from '@nestjs/jwt';
+import { JwtAuthGuard } from 'src/modules/usersModule/auth/application/jwt.auth.guard';
 
 class UpdateCommentInputDto {
   @IsNotEmpty()
