+++ nest_lesson_3/src/modules/commonModule/common.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+
+@Module({
+  imports: [],
+  providers: [],
+  exports: [],
+})
+export class CommonModule {}
