+++ nest_lesson_3/src/modules/usersModule/auth/application/auth.jwt.strategy.ts
