+++ nest_lesson_3/src/modules/usersModule/auth/application/jwt.auth.guard.ts
