+++ nest_lesson_3/src/modules/usersModule/users/domain/user.entity.ts
